feat : export 구현중, env key, value 추가#36
