+++ kipark_jayoon_minishell/includes/built.h
@@ -6,7 +6,7 @@
 /*   By: kipark <kipark@student.42seoul.kr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/16 17:32:29 by kipark            #+#    #+#             */
-/*   Updated: 2022/09/22 16:37:54 by kipark           ###   ########seoul.kr  */
+/*   Updated: 2022/09/22 22:26:12 by kipark           ###   ########seoul.kr  */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,16 +14,6 @@
 # define BUILT
 
 #include"parser.h"
-
-#define CD_SIZE 2
-#define EHCO_SIZE 4
-#define ENV_SIZE 3
-#define EXIT_SIZE 4
-#define EXPORT_SIZE 6
-#define PWD_SIZE 3
-#define UNSET_SIZE 5
-
-
 
 int		is_built_in(t_simple_cmd *simple_cmd);
 void	execute_bulit_in(t_simple_cmd *simple_cmd, t_env *env);
