+++ kipark_jayoon_minishell/includes/env.h
@@ -6,7 +6,7 @@
 /*   By: kipark <kipark@student.42seoul.kr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/26 13:21:31 by kipark            #+#    #+#             */
-/*   Updated: 2022/08/30 14:39:38 by kipark           ###   ########seoul.kr  */
+/*   Updated: 2022/09/22 21:13:43 by kipark           ###   ########seoul.kr  */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,12 +16,15 @@
 typedef struct s_env
 {
 	char			*str;
+	char			*key;
+	char			*value;
 	struct s_env	*next;
 }	t_env;
 
 
 // env_util*.c
 char		*get_env_value(t_env *env_head, char *env_key);
+int			get_env_equal_location(char *env_str);
 int			get_env_key_size(char *env_key);
 t_env		*set_shell_env_list(char **envp);
