+++ kipark_jayoon_minishell/srcs/built_in/built_in.c
@@ -6,7 +6,7 @@
 /*   By: kipark <kipark@student.42seoul.kr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/17 13:26:41 by kipark            #+#    #+#             */
-/*   Updated: 2022/09/22 17:26:07 by kipark           ###   ########seoul.kr  */
+/*   Updated: 2022/09/22 22:26:14 by kipark           ###   ########seoul.kr  */
 /*                                                                            */
 /* ************************************************************************** */
