+++ kipark_jayoon_minishell/srcs/built_in/env.c
@@ -6,7 +6,7 @@
 /*   By: kipark <kipark@student.42seoul.kr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/17 12:46:51 by kipark            #+#    #+#             */
-/*   Updated: 2022/09/22 16:30:26 by kipark           ###   ########seoul.kr  */
+/*   Updated: 2022/09/22 21:26:34 by kipark           ###   ########seoul.kr  */
 /*                                                                            */
 /* ************************************************************************** */
