+++ kipark_jayoon_minishell/srcs/terminal/free_terminal.c
@@ -6,7 +6,7 @@
 /*   By: kipark <kipark@student.42seoul.kr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/02 11:36:24 by kipark            #+#    #+#             */
-/*   Updated: 2022/09/22 17:00:35 by kipark           ###   ########seoul.kr  */
+/*   Updated: 2022/09/22 22:03:48 by kipark           ###   ########seoul.kr  */
 /*                                                                            */
 /* ************************************************************************** */
